Shiny Dashboard for Sample Size and Power Full Draft
The entire application and documentation, except the statistical
equations, have been reviewed and edited. Review and edits to the
equations are to come. The application is completely working and all
known bugs have been fixed. The documentation has a .Rmd raw version as
well as a .docx and .html complete versions.